Allow Alarm to take a sensor as an argument
By allowing Alarm to take a sensor as an argument it allows us to test
the class more easily. In our test we can now control what `sensor` is;
so we have passed in an `instance_double` where we can dictate the
values returned by the methods called on it.

It also satifies the O in SOLID more as if we ever want to use a
different sensor we just have to pass it into Alarms initilizer instead
of editing the Alarm class.