Cleaned up the defaulted attr_reader as it's not necessary
As the `README.md` states:
  ```Take care when refactoring not to alter the functionality, or change
  interfaces which other client code may rely on```

We don't want to alter the public API of the Alarm class as other code
may be dependant on it i.e. being able to call the `check` method and
access the `@is_alarm_on` attribute. The `check` method is all fine and
dandy as another class will be able to call it, but we have removed the
`@is_alarm_on` instance variable along with it's `attr_reader`. However
we can still maintain the public API by creating an alias method
`is_alarm_on` which will execute the `check` method.

Before this refactor another class would have been able to do:
```
  a = Alarm.new(Sensor.new(), (17..21))
  a.is_alarm_on # => false because that was what the default was
  a.check       # => true or false depending on what randomness Sensor used
  a.is_alarm_on # => whatever a.check was
```
and now it can still call all those methods in that order and get the
correct outcome for whether the alarm is on or not